add support to update datasources
So far this module has not been able to update datasource to target another server. This is useful for me.
I added some more detailed error handling because it is not obvious that a 404 does not mean that the URL does not exist but can as well mean that the selector does no match any current data source.