🐛 Not working when table expression contained reference to another table
When M expression had reference to another table (ex Merge), the old code would create a new ParameterExpression and blow up when the table with the same name was created.
To solve this I've added a guard to check if the object is a hidden table before the Expression would be created.